Add files via upload
In this case study, I have walked through the process of setting up a Jupyter Notebook with NumPy and provided examples of how to use the library for scientific computing. Jupyter Notebook is a powerful tool that allows you to prototype and experiment with code in a flexible and interactive environment. NumPy is a popular library for scientific computing in Python that provides support for arrays and matrices, as well as a range of mathematical functions that can be performed on these arrays. With these tools, I can easily analyze data, build machine learning models, and perform scientific calculations in Python.